Extract `PostgreSQL::SchemaCreation` to appropriate file
As below:

https://github.com/rails/rails/blob/master/activerecord/lib/active_record/connection_adapters/mysql/schema_creation.rb
https://github.com/rails/rails/blob/master/activerecord/lib/active_record/connection_adapters/sqlite3/schema_creation.rb
